Create FCTRL - Factorial
Uses a special property ...only 5 and 2 contribute for trailing zeroes in a factorial.
Also, note that no of 5 always <=2. So,its sufficient to count number of 5's in factorial of a number.
Also note that 25 (5*5) has two fives. 125 has three 5's.
So,just simply get n/5 + n/25 + n/125 ... and so on.